refactor: update import paths and add TODO comments
The changes include:
- Update import paths from 'use-role' to 'useRole'
- Add TODO comments for pending tasks in timer and webhook routes
- Add file path comments at the top of files
- Add invitation types export in types/index.ts
