+++ src/app/(auth)/clerk_invitations/actions/getInvitations.ts
@@ -1,5 +1,5 @@
 // src/app/actions/getInvitations.ts
-// DONE fix type regarding api (specifically expires_at )
+//DONE: fix type regarding api (specifically expires_at )
 'use server'
 
 import { clerkClient } from '@clerk/nextjs/server'
