+++ src/app/components/ui/toaster.tsx
@@ -1,6 +1,7 @@
+// src/app/components/ui/toaster.tsx
 "use client"
 
-import { useToast } from "@/hooks/use-toast"
+import { useToast } from "@/hooks/useToast"
 import {
   Toast,
   ToastClose,
