+++ src/hooks/useTimer.ts
@@ -1,3 +1,4 @@
+// src/hooks/useTimer.ts
 import { useState, useCallback, useRef, useEffect } from 'react';
 
 export const useTimer = () => {
