+++ src/hooks/useToast.ts
@@ -1,5 +1,5 @@
 "use client"
-
+// src/hooks/use-toast.ts
 // Inspired by react-hot-toast library
 import * as React from "react"
