+++ src/middleware.ts
@@ -1,3 +1,4 @@
+//src/middleware.ts
 import { clerkMiddleware, createRouteMatcher } from '@clerk/nextjs/server'
 
 const isProtectedRoutes = createRouteMatcher([
