+++ src/types/index.ts
@@ -1,2 +1,4 @@
+//src/types/index.ts
 // Central type exports
 export * from './user'
+export * from './invitation'
